moved changes to be for all navigatin and not just track.trackOnce
